chore: rename rule name
