+++ lib/configs/no-layout-rules.js
@@ -41,8 +41,8 @@ module.exports = {
     'vue/object-property-newline': 'off',
     'vue/operator-linebreak': 'off',
     'vue/padding-line-between-blocks': 'off',
-    'vue/prefer-type-emits-decl': 'off',
-    'vue/prefer-type-props-decl': 'off',
+    'vue/define-emits-declaration': 'off',
+    'vue/define-props-declaration': 'off',
     'vue/script-indent': 'off',
     'vue/singleline-html-element-content-newline': 'off',
     'vue/space-in-parens': 'off',
