+++ lib/index.js
@@ -162,8 +162,8 @@ module.exports = {
     'prefer-separate-static-class': require('./rules/prefer-separate-static-class'),
     'prefer-template': require('./rules/prefer-template'),
     'prefer-true-attribute-shorthand': require('./rules/prefer-true-attribute-shorthand'),
-    'prefer-type-emits-decl': require('./rules/prefer-type-emits-decl'),
-    'prefer-type-props-decl': require('./rules/prefer-type-props-decl'),
+    'define-emits-declaration': require('./rules/define-emits-declaration'),
+    'define-props-declaration': require('./rules/define-props-declaration'),
     'prop-name-casing': require('./rules/prop-name-casing'),
     'quote-props': require('./rules/quote-props'),
     'require-component-is': require('./rules/require-component-is'),
