+++ lib/rules/define-emits-declaration.js
@@ -20,7 +20,7 @@ module.exports = {
     docs: {
       description: 'enforce type-based `defineEmits`',
       categories: undefined,
-      url: 'https://eslint.vuejs.org/rules/prefer-type-emits-decl.html'
+      url: 'https://eslint.vuejs.org/rules/define-emits-declaration.html'
     },
     fixable: null,
     schema: [],
