+++ lib/rules/define-props-declaration.js
@@ -24,7 +24,7 @@ module.exports = {
     docs: {
       description: 'enforce type-based `defineProps`',
       categories: undefined,
-      url: 'https://eslint.vuejs.org/rules/prefer-type-props-decl.html'
+      url: 'https://eslint.vuejs.org/rules/define-props-declaration.html'
     },
     fixable: null,
     schema: [],
