+++ tests/lib/rules/define-props-declaration.js
@@ -5,7 +5,7 @@
 'use strict'
 
 const RuleTester = require('eslint').RuleTester
-const rule = require('../../../lib/rules/prefer-type-emits-decl')
+const rule = require('../../../lib/rules/define-props-declaration')
 
 const tester = new RuleTester({
   parser: require.resolve('vue-eslint-parser'),
@@ -15,26 +15,27 @@ const tester = new RuleTester({
   }
 })
 
-tester.run('prefer-type-emits-decl', rule, {
+tester.run('define-props-declaration', rule, {
   valid: [
     {
       filename: 'test.vue',
       code: `
-        <script setup>
-          const emit = defineEmits(['change', 'update'])
-        </script>
-       `
+      <script setup>
+      const props = defineProps({
+        kind: { type: String },
+      })
+      </script>
+      `
     },
     {
       filename: 'test.vue',
       code: `
-        <script setup lang="ts">
-        const emit = defineEmits<{
-          (e: 'change', id: number): void
-          (e: 'update', value: string): void
-        }>()
-        </script>
-       `,
+      <script setup lang="ts">
+      const props = defineProps<{
+        kind: string;
+      }>()
+      </script>
+      `,
       parserOptions: {
         parser: require.resolve('@typescript-eslint/parser')
       }
@@ -42,12 +43,12 @@ tester.run('prefer-type-emits-decl', rule, {
     {
       filename: 'test.vue',
       code: `
-        <script setup lang="ts">
-        const props = defineProps({
-          kind: { type: String },
-        })
-        </script>
-       `,
+      <script setup lang="ts">
+      const emit = defineEmits({
+        click: (event: PointerEvent) => !!event
+      })
+      </script>
+      `,
       parserOptions: {
         parser: require.resolve('@typescript-eslint/parser')
       }
@@ -55,17 +56,17 @@ tester.run('prefer-type-emits-decl', rule, {
     {
       filename: 'test.vue',
       code: `
-         <script lang="ts">
-         import { PropType } from 'vue'
- 
-         export default {
-           props: {
-             kind: { type: String as PropType<'primary' | 'secondary'> },
-           },
-           emits: ['check']
-         }
-         </script>
-       `,
+        <script lang="ts">
+        import { PropType } from 'vue'
+
+        export default {
+          props: {
+            kind: { type: String as PropType<'primary' | 'secondary'> },
+          },
+          emits: ['check']
+        }
+        </script>
+      `,
       parserOptions: {
         parser: require.resolve('@typescript-eslint/parser')
       }
@@ -75,10 +76,12 @@ tester.run('prefer-type-emits-decl', rule, {
     {
       filename: 'test.vue',
       code: `
-       <script setup lang="ts">
-       const emit = defineEmits(['change', 'update'])
-       </script>
-       `,
+      <script setup lang="ts">
+      const props = defineProps({
+        kind: { type: String },
+      })
+      </script>
+      `,
       errors: [
         {
           message: 'Use type-based declaration instead of runtime declaration.',
